apps/qaqc/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./bash/compile_python_packages.sh
#
altair==5.0.1
    # via streamlit
attrs==23.1.0
    # via
    #   fiona
    #   jsonschema
blinker==1.6.2
    # via streamlit
boto3==1.24.47
    # via -r requirements.in
botocore==1.27.96
    # via
    #   boto3
    #   s3transfer
cachetools==5.3.1
    # via streamlit
certifi==2023.5.7
    # via
    #   fiona
    #   pyproj
    #   requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   streamlit
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
decorator==5.1.1
    # via validators
fiona==1.9.4.post1
    # via geopandas
fonttools==4.40.0
    # via matplotlib
geopandas==0.13.2
    # via -r requirements.in
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via requests
importlib-metadata==6.7.0
    # via streamlit
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via altair
kiwisolver==1.4.4
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
matplotlib==3.7.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
numerize==0.12
    # via -r requirements.in
numpy==1.25.0
    # via
    #   -r requirements.in
    #   altair
    #   contourpy
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   shapely
    #   streamlit
packaging==23.1
    # via
    #   geopandas
    #   matplotlib
    #   plotly
    #   pytest
    #   streamlit
pandas==1.5.3
    # via
    #   -r requirements.in
    #   altair
    #   geopandas
    #   streamlit
    #   streamlit-aggrid
pillow==9.5.0
    # via
    #   matplotlib
    #   streamlit
plotly==5.15.0
    # via -r requirements.in
pluggy==1.2.0
    # via pytest
protobuf==4.23.3
    # via streamlit
psycopg2-binary==2.9.6
    # via -r requirements.in
pyarrow==12.0.1
    # via streamlit
pydeck==0.8.0
    # via streamlit
pygments==2.15.1
    # via rich
pympler==1.0.1
    # via streamlit
pyparsing==3.1.0
    # via matplotlib
pyproj==3.6.0
    # via geopandas
pyrsistent==0.19.3
    # via jsonschema
pytest==7.3.0
    # via -r requirements.in
python-abc==0.2.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   matplotlib
    #   pandas
    #   streamlit
python-decouple==3.8
    # via streamlit-aggrid
python-dotenv==0.21.1
    # via -r requirements.in
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.31.0
    # via streamlit
rich==13.4.2
    # via streamlit
s3transfer==0.6.1
    # via boto3
shapely==2.0.1
    # via geopandas
six==1.16.0
    # via
    #   fiona
    #   python-dateutil
smmap==5.0.0
    # via gitdb
sqlalchemy==2.0.9
    # via -r requirements.in
streamlit==1.24.0
    # via
    #   -r requirements.in
    #   streamlit-aggrid
streamlit-aggrid==0.3.4.post3
    # via -r requirements.in
tabulate==0.9.0
    # via -r requirements.in
tenacity==8.2.2
    # via
    #   plotly
    #   streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.2
    # via streamlit
typing-extensions==4.6.3
    # via
    #   sqlalchemy
    #   streamlit
tzdata==2023.3
    # via pytz-deprecation-shim
tzlocal==4.3.1
    # via streamlit
urllib3==1.26.16
    # via
    #   botocore
    #   requests
validators==0.20.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
zipp==3.15.0
    # via importlib-metadata
